all: fix warnings
cpu.c:587:4: warning: ‘n’ may be used uninitialized in this function
cpu.c:659:3: warning: dereferencing type-punned pointer will break strict-aliasing rules
cpu.c:659:3: warning: dereferencing type-punned pointer will break strict-aliasing rules
kern/devaudio.c:238:25: warning: variable ‘in’ set but not used
kern/devaudio.c:240:8: warning: variable ‘a’ set but not used
kern/sysproc.c:14:11: warning: cast to pointer from integer of different size
libc/nsec.c:56:12: warning: implicit declaration of function ‘libstrtoll’
libc/runestrchr.c:16:2: warning: suggest parentheses around assignment used as truth value
libmemdraw/draw.c:2209:4: warning: dereferencing type-punned pointer will break strict-aliasing rules
libmemdraw/draw.c:2224:4: warning: dereferencing type-punned pointer will break strict-aliasing rules
libmemdraw/draw.c:2455:4: warning: dereferencing type-punned pointer will break strict-aliasing rules
libmemdraw/draw.c:2481:4: warning: dereferencing type-punned pointer will break strict-aliasing rules

+++ kern/sysproc.c
@@ -5,7 +5,7 @@
 #include	"error.h"
 
 long
-sysexits(ulong *arg)
+sysexits(uintptr_t *arg)
 {
 	char *status;
 	char *inval = "invalid exit string";
